// server/src/enums.ts
export enum Category {
  LESSON = "lesson",
  DRILL = "drill",
  QUOTE = "quote",
  GAME = "game"
}

export enum DifficultyLevel {
  BEGINNER = 'beginner',
  INTERMEDIATE = 'intermediate',
  EXPERT = 'expert'
}

export enum Language {
  ENGLISH = "english",
  FRENCH = "french",
  SPANISH = "spanish",
  GERMAN = "german"
}


export enum UserProgressStatus {
  LOCKED = "locked",
  AVAILABLE = "available",
  COMPLETED = "completed"
}

export enum ThemeSetting {
  DARK = "dark",
  LIGHT = "light"
}

export enum KeyboardSetting {
  ON = "on",
  OFF = "off"
}